01-mcp-study/4_chromaDB/pyproject.toml
[project]
name = "4-chromaDB"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chromadb>=1.0.15",
    "ddgs>=9.5.1",
    "google-genai>=1.28.0",
    "logger>=1.4",
    "mcp[cli]>=1.12.3",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.0.0",
]
